chatbot/requirements.txt
google-genai
python-dotenv
chromadb
flask
flask-cors
PyPDF2
python-docx
openpyxl
python-pptx